Add Fennelucky creation functionality
New functionality has been added for creating a Fennelucky. This includes updates to the Fennelucky model to include a user relationship and the added fillable properties 'user_id' and 'name'. Additionally, a new FenneluckyController is created with a store method for handling the creation request, and a new POST api route for fennelucky creation has been created. This is to allow users to create their own Fenneluckies.